Main activity code cleanup and documentation
