+++ app/src/main/AndroidManifest.xml
@@ -22,14 +22,14 @@
             android:noHistory="false"/>
         <activity android:name=".activity.ListActivity"
             />
-        <activity android:name=".activity.NewMainActivity">
+        <activity android:name=".activity.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".activity.MainActivity"
+        <activity android:name=".activity.OldMainActivity"
             />
 
 
@@ -52,7 +52,7 @@
         <activity android:name=".activity.VoiceToTextActivity"
             />
 
-        <activity android:name=".activity.XTemplateEmptyActivity"
+        <activity android:name=".test.XTemplateEmptyActivity"
             />
 
         <activity android:name=".test.DebugActivity"
