+++ app/src/main/res/layout/activity_voice_to_text.xml
@@ -22,7 +22,7 @@
             app:popupTheme="@style/ThemeOverlay.AppCompat.Light" />
 
         <TextView
-            android:id="@+id/template_textView"
+            android:id="@+id/template_text_view"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:text="VOICE"
