+++ app/src/main/res/menu/drawer_menu.xml
@@ -21,7 +21,7 @@
         <item
             android:id="@+id/drawer_debug"
             android:title="Debug"
-            android:visible="false"
+            android:visible="true"
              />
     </group>
